microservices/api_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
aiokafka
prometheus-fastapi-instrumentator
authlib
uvicorn
PyJWT>=2.8.0
cryptography